+++ __mocks__/asset-string.js
@@ -1,0 +1,1 @@
+module.exports = "data:application/wasm;base64,AAAA";