+++ jest.config.js
@@ -89,7 +89,10 @@ module.exports = {
   ],
 
   // A map from regular expressions to module names or to arrays of module names that allow to stub out resources with a single module
-  // moduleNameMapper: {},
+  moduleNameMapper: {
+    "\\.wasm$": "<rootDir>/__mocks__/asset-string.js",
+    "\\?worker-source$": "<rootDir>/__mocks__/asset-string.js",
+  },
 
   // An array of regexp pattern strings, matched against all module paths before considered 'visible' to the module loader
   // modulePathIgnorePatterns: [],