+++ src/types/assets.d.ts
@@ -1,0 +1,9 @@
+declare module "*.wasm" {
+    const source: string;
+    export default source;
+}
+
+declare module "*?worker-source" {
+    const source: string;
+    export default source;
+}